Crate collatz_sequence

This algorithm will compute the length of the Collatz sequence for all numbers in a given range. It will then output:

The number in the provided range with the longest Collatz sequence.

The biggest number computed while computing the Collatz sequence.

Collatz Sequence

The Collatz sequence is defined as follows:

If the number is even, divide it by 2.

If the number is odd, multiply it by 3 and add 1. And you repeat this process until you reach 1. The length of the sequence is the number of steps needed to reach 1.

The Collatz conjecture is the conjecture that for any number, the Collatz sequence will always reach 1. Hypothetically, the sequence could go on forever, or find another cycle, as this conjecture has not yet been proven.

Functions

collatz_sequence 🔒

Compute the collatz sequence for a given number and add every step to the cache. This function will use the cache to avoid computing the same sequence twice.

collatz_step 🔒

Compute the next number in the collatz sequence. We use a BigUint because the numbers can get very big very quickly.

main 🔒